+++ packages/adapter-interfaces/src/types/LoginSession.ts
@@ -10,7 +10,7 @@ export const loginSessionInsertSchema = z
     deleted_at: z.string().optional(),
     ip: z.string().optional(),
     useragent: z.string().optional(),
-    session: z.string().optional(),
+    session_id: z.string().optional(),
     authorization_url: z.string().optional(),
   })
   .openapi({